feat: Implement project sorting utility for homepage
* Added a new utility function `sortHomepageProjects` to sort project objects based on a predefined order and return the top 3 projects.
* Updated the `Home` component to utilize the new sorting function for experiments data, enhancing project display logic.
* Data size on the page is reduced
